requirements.txt
openai
wavio
sounddevice
scipy
python-dotenv
streamlit
streamlit_chat
flask
os
json
